Add support for building in Docker
This commit introduces an alternative builder for Debian that
uses a Docker container rather than a VirtualBox VM. In many cases,
this results in significantly faster builds, in particular due to the
ability to avoid rsyncing large numbers of files. Additionally, the
overhead of running the VirtualBox VM can be avoided.

NOTE: This does not _replace_ any builders, but rather adds support
for building with either Docker or VirtualBox, depending on preference.

By default the builds will continue to use VirtualBox, but you can
choose Docker by using the make tasks set up for that purpose in the
root Makefile:

    # Build the base Docker image, equivalent to creating the base VirtualBox image
    make builder

    # Once built, you can use one of the following:
    make docker

    # OR
    make virtualbox

You can skip the use of make for builds and just export
`VAGRANT_DEFAULT_PROVIDER=docker` instead, if that's your preference.